% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractr.R
\name{abstractr}
\alias{abstractr}
\title{Launch abstractr GUI}
\usage{
abstractr()
}
\description{
abstractr() loads interactive user interface built using R shiny.
}
\details{
Start abstractr

The interactive application makes creating visual abstracts fast and intuitive. While the application has certain design elements fixed, namely layout, a vignette is included to guide customizations from scratch.
}
\examples{
\dontrun{
library(shiny)
abstractr()
}
}
\keyword{abstractr}
